change default downloading behavior; mist 1.1 default
This enables versioning the MIST model grids, and makes v1.1 default.
It also changes the default download behavior for both Dartmouth and
MIST, such that there is no longer a master tarball of all the grids to
download; rather, individual grid tarballs are downloaded as needed.

+++ isochrones/utils.py
@@ -3,11 +3,12 @@ from __future__ import print_function, division
 from .config import on_rtd
 import requests
 import os
+import logging
 
 if not on_rtd:
     import numpy as np
 
-def download_file(url, path=None):
+def download_file(url, path=None, clobber=False):
     """
     thanks to: https://stackoverflow.com/questions/16694907/how-to-download-large-file-in-python-with-requests-py
 
@@ -18,11 +19,15 @@ def download_file(url, path=None):
         local_filename = os.path.join(directory, url.split('/')[-1])
     else:
         local_filename = path
+
+    if os.path.exists(local_filename) and not clobber:
+        logging.info('{} exists; not downloading.'.format(local_filename))
+        return local_filename
         
     # NOTE the stream=True parameter
     r = requests.get(url, stream=True)
     with open(local_filename, 'wb') as f:
-        for chunk in r.iter_content(chunk_size=1024): 
+        for chunk in r.iter_content(chunk_size=1024):
             if chunk: # filter out keep-alive new chunks
                 f.write(chunk)
                 #f.flush() commented by recommendation from J.F.Sebastian
@@ -44,15 +49,15 @@ def addmags(*mags):
             tot += f
             unc = f * (1 - 10**(-0.4*dm))
             uncs.append(unc)
-    
+
     totmag = -2.5*np.log10(tot)
     if len(uncs) > 0:
         f_unc = np.sqrt(np.array([u**2 for u in uncs]).sum())
         return totmag, -2.5*np.log10(1 - f_unc/tot)
     else:
-        return totmag 
-    
-    
+        return totmag
+
+
 def distance(pos0, pos1):
     """distance between two positions defined by (separation, PA)
     """
@@ -60,7 +65,7 @@ def distance(pos0, pos1):
     #logging.debug('r0={}, pa0={} (from {})'.format(r0, pa0, self))
     ra0 = r0*np.sin(pa0*np.pi/180)
     dec0 = r0*np.cos(pa0*np.pi/180)
-    
+
     r1, pa1 = pos1
     #logging.debug('r1={}, pa1={} (from {})'.format(r0, pa0, other))
     ra1 = r1*np.sin(pa1*np.pi/180)
